Make Transaction fetching behavior pluggable
